third_party/blink/web_tests/wpt_internal/fenced_frame/unpartitioned-storage-permissions-policy-allow.tentative.https.sub.html

<!DOCTYPE html>
<title>Test fenced-unpartitioned-storage-read permissions policy with "allow" attribute</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
<script src="resources/unpartitioned_storage_permissions_policy_helper.js"></script>

<body>
<script>

promise_setup(async () => {
  await configureSharedStorageDataForTesting();
});


promise_test(async (t) => {
  let get_result = await runUnpartitionedStoragePermissionsPolicyTestCase(
      get_host_info().HTTPS_REMOTE_ORIGIN,
     'allow', get_host_info().HTTPS_REMOTE_ORIGIN);
  assert_equals(get_result, 'banana');
}, 'Test that a fenced-unpartitioned-storage-read allowlist set via allow ' +
   'attribute enables calling of sharedStorage.get()');


promise_test(async (t) => {
  let get_result = await runUnpartitionedStoragePermissionsPolicyTestCase(
      get_host_info().HTTPS_ORIGIN,
     'allow', get_host_info().HTTPS_REMOTE_ORIGIN);
  assert_equals(get_result, 'Permission denied');
}, 'Test that a fenced-unpartitioned-storage-read allowlist set via allow ' +
   'attribute properly restricts calling of sharedStorage.get() from a ' +
   'disallowed origin');

</script>
</body>
